% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Data}
\alias{Data}
\title{Nahraj data}
\usage{
Data(...)
}
\arguments{
\item{...}{jméno/jména datasetu}
}
\description{
Umožňuje přistupovat k datasetům ze synchronizované owncloud složky "used_data". Většina atributů načítá rovnou více datasetů. Stručný popis, viz níže. Pro přidání a dokumentaci datasetu upravte kód na githubu, nasynchronizujte data na owncloudu a znovu sestavte balík. Nebo kontaktujte správce.
}
\section{Data(chars)}{


\strong{CHARS} (data.table): Hydrologické charakteristiky povodi UPOV
\itemize{
\item UPOV_ID
\item NAZ_UTVAR Název útvaru
}
}

\section{Data(geo)}{


\strong{NADRZE} (shapefile): Významné vodní nádrže
\itemize{
\item položky atributové tabulky zatím nejasné
}

\strong{UPOVS} (shapefile): Vrstva povodí
\itemize{
\item UPOV_ID
}
}

\section{Data(qd)}{


\strong{QD} (data.table): Denní průtoky
}

\section{Data(routing)}{


\strong{TABA} (data.table): Tabulka následnosti povodí
\itemize{
\item FROM: UPOV_ID odkud (tece voda)
\item TO: UPOV_ID kam (tece voda), -9999 = mimo republiku
\item FR_ORD, TO_ORD: rady toku
\item UPOV_AREA: plocha povodi - jen UPOV
\item TOTAL_AREA: celkova plocha povodi - az k prameni - v soucasnosti pro povodi Ohre a ostatni nedodana povodi dopocteno jako suma UPOV_AREA - nekde to muze byt uplne blbe
}

\strong{TABB} (data.table): Tabulka vsech hornich povodi pro danny UPOV
\itemize{
\item FROM:
\item TO:
}

\strong{TABNAD} (data.table): Tabulka charakteristik nadrží

\strong{TABREKY} (data.table): Tabulka delek rek pro HYPE v m
}

\section{Data(uziv06)}{


\strong{UZIV06} (data.table): užívání 2006-2016 v tis.m3/mes
\itemize{
\item ?
}

\strong{UZIV06mm} (data.table): užívání 2006-2016 v mm/mes
\itemize{
\item ?
}
}

\section{Data(uziv79)}{


\strong{UZIV79} (data.table): užívání 1979-2016 v tis.m3/mes
\itemize{
\item ?
}

\strong{UZIV79mm} (data.table): užívání 1979-2016 v mm/mes
\itemize{
\item ?
}
}

\section{Data(webapp_data)}{


\strong{JEZERA}
\itemize{
\item ?
}

\strong{POVODI}
\itemize{
\item ?
}

\strong{REKY}
\itemize{
\item ?
}

\strong{STANICE}
\itemize{
\item ?
}
}

