% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{give_paths}
\alias{give_paths}
\title{Načte globální cesty}
\usage{
give_paths
}
\value{
Přiřadí do \code{.GlobalEnv} proměnné \code{.datadir} a \code{.workdir}
}
\description{
Načte proměnné \code{.datadir} (synchronizovaná owncloud složka "used_data" z BILAN_UPOV) a \code{.workdir} (zatím nevyužito) na základě jména počítače. Před použitím je nutné PC registrovat - buď úpravou kódu na githubu nebo emailem na \email{hanel@fzp.czu.cz} ve struktuře
\preformatted{
jmeno_PC = { 
  .datadir = 'cesta do used_data' 
  .workdir = 'pracovni cesta' 
} 
}
}
